Add files via upload
We used this script to carry out the analysis on the italian neuroscience articles, sorting these by macroregions (Northern Italy, Central Italy, and Southern Italy and Islands), university type (public or private) and private university type (non-telematic and telematic). This script can also be used to find the average index from a generic distribution of articles. To find out what each section stands for, read the comments inside the file.